requirements.txt
streamlit>=1.33.0
pandas>=2.1.0
matplotlib>=3.8.0
openpyxl>=3.1.0
zhipuai>=2.0.0
python-dotenv>=1.0.0

